pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "casesentinel"
version = "0.1.0"
description = "LLM-native multi-agent research platform for criminal investigation."
readme = "README.md"
authors = [{ name = "CaseSentinel Team" }]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "jinja2>=3.1.0",
    "pydantic>=2.7.0",
    "tqdm>=4.66.0",
    "openai>=1.23.6",
    "python-dotenv>=1.0.0",
    "langchain>=0.2.15",
    "langchain-community>=0.2.15",
    "chromadb>=0.4.24",
    "networkx>=3.3",
    "uvicorn>=0.29.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["src"]
